Added a Windows .bat file - starts servers
Added a windows .bat file which starts the modules for Bassa. This allows for extra user compatibility with the windows operating system. Before starting the modules the bat file checks for dependencies. Currently it checks for python, node, npm, and gulp. If these are not installed it will produce an error. If they are installed it will continue and start the servers