Made the script Docker-compatible + Prevented version failures
## Made setup.sh compatible for Docker and linux user:
Docker always runs as root, Earlier the "sudo" commands were making trouble for Docker but not for linux users. Docker can't access the ```sudo``` command as its $USER is already set as root. In this commit, I've removed all the ```sudo``` (s) in the shell script. OK, but wouldn't this mess up the installation for linux users since you require some permission to install packages. To take care of that, I added an If-else condition of detecting the $USER, If that's found as a "non-root" user, An echo of instructions to proceed with root user is given out in the terminal, When linux users switch to SuperUser (su) and execute this script, the installation happens flawlessly.

## Prevented version failures
It's important to first update package list, then update the versions of all pre-installed packages in the system. This script installs a few popular things like npm,pip,nodejs etc.. 
Many people like me are already having them. In that case, APT just ignore the installation if the presence of that specific package is detected, and this results in the operations of older versions of packages, which to an extent may cause depreciation warnings or failures. If we update and upgrade the versions of the packages, The build gets cleaner and safer...