Made the script Docker-compatible + Prevented version failures
Docker always runs as root, Earlier the "sudo" commands were making trouble for Docker but not for bash users. Docker can't access the sudo command as its $USER is already set as root. In this commit, I've removed all the sudo (s) in the shell script while ensuring support for docker users...